Add a keepalive value + use constants for headers
