+++ providers/go-feature-flag/pkg/controller/http_client.go
@@ -8,6 +8,7 @@ import (
 func DefaultHTTPClient() *http.Client {
 	netTransport := &http.Transport{
 		TLSHandshakeTimeout: 10000 * time.Millisecond,
+		IdleConnTimeout:     90 * time.Second,
 	}
 
 	return &http.Client{
