+++ providers/go-feature-flag/pkg/controller/http_constants.go
@@ -1,0 +1,8 @@
+package controller
+
+const ContentTypeHeader = "Content-Type"
+const IfNoneMatchHeader = "If-None-Match"
+const AuthorizationHeader = "Authorization"
+
+const ApplicationJson = "application/json"
+const BearerPrefix = "Bearer "
